# Total quantity cache column for Itemizables