--- a/Gemfile
+++ b/Gemfile
@@ -15,6 +15,7 @@ gem 'bootstrap-select-rails'
 gem "bugsnag"
 gem "chartkick"
 gem "cocoon"
+gem "counter_culture", "~> 2.8"
 gem "devise", '>= 4.7.1'
 gem 'discard', '~> 1.2'
 gem "devise_invitable"