--- a/app/pdfs/distribution_pdf.rb
+++ b/app/pdfs/distribution_pdf.rb
@@ -51,7 +51,7 @@ class DistributionPdf
     data += @distribution.line_items.sorted.map do |c|
       [c.item.name, dollar_value(c.item.value_in_cents), dollar_value(c.value_per_line_item), c.quantity, c.package_count]
     end
-    data += [["", "", "", "", ""], ["Total Items Received", "", dollar_value(@distribution.value_per_itemizable), @distribution.line_items.total, ""]]
+    data += [["", "", "", "", ""], ["Total Items Received", "", dollar_value(@distribution.value_per_itemizable), @distribution.line_items_total, ""]]
 
     font_size 11
     # Line item table