--- a/app/views/dashboard/_diaper_drive.html.erb
+++ b/app/views/dashboard/_diaper_drive.html.erb
@@ -1,5 +1,5 @@
   <div class="float-center">
     <%= link_to donation do %>
-      <%= donation.line_items.total %> from <%= donation.diaper_drive&.name || donation.diaper_drive_participant.business_name %>
+      <%= donation.line_items_total %> from <%= donation.diaper_drive&.name || donation.diaper_drive_participant.business_name %>
     <% end %>
   </div>