--- a/app/views/dashboard/_distribution.html.erb
+++ b/app/views/dashboard/_distribution.html.erb
@@ -1,7 +1,7 @@
 <br>
 <div class="pull-left">
   <%= link_to distribution do %>
-    <%= distribution.line_items.total %> items distributed to
+    <%= distribution.line_items_total %> items distributed to
     <%= distribution.partner_name %>
   <% end %>
 </div>