--- a/app/views/dashboard/_donation.html.erb
+++ b/app/views/dashboard/_donation.html.erb
@@ -1,7 +1,7 @@
 <br>
 <div class="pull-left">
   <%= link_to donation do %>
-    <%= donation.line_items.total %> items from <%= donation.source %>
+    <%= donation.line_items_total %> items from <%= donation.source %>
   <% end %>
 </div>
 <div class="pull-right">