--- a/app/views/dashboard/_purchase.html.erb
+++ b/app/views/dashboard/_purchase.html.erb
@@ -1,7 +1,7 @@
 <br>
 <div class="pull-left">
   <%= link_to purchase do %>
-    <%= purchase.line_items.total %> items from <%= purchase.purchased_from_view %>
+    <%= purchase.line_items_total %> items from <%= purchase.purchased_from_view %>
   <% end %>
 </div>
 <div class="pull-right">