--- a/app/views/donations/_donation_row.html.erb
+++ b/app/views/donations/_donation_row.html.erb
@@ -3,7 +3,7 @@
   <td class="text-left date"><%= donation_row.issued_at.strftime("%F") %>
   <td><%= donation_row.donation_site.try(:name) %></td>
   <td><%= donation_row.storage_location.name %></td>
-  <td class="text-left numeric"><%= donation_row.line_items.total %></td>
+  <td class="text-left numeric"><%= donation_row.line_items_total %></td>
   <td class="text-left numeric"><%= dollar_value(donation_row.money_raised.to_i) %></td>
   <td class="text-left numeric"><%= dollar_value(donation_row.value_per_itemizable) %></td>
   <td><%= truncate donation_row.comment, length: 140, separator: /\w+/ %>