--- a/app/views/manufacturers/show.html.erb
+++ b/app/views/manufacturers/show.html.erb
@@ -46,7 +46,7 @@
               <% @manufacturer.donations.each do |donation| %>
                 <tr>
                   <td><%= donation.created_at.strftime("%m/%d/%Y") %></td>
-                  <td><%= donation.line_items.total %></td>
+                  <td><%= donation.line_items_total %></td>
                   <td>
                     <%= view_button_to donation, { text: "View donation details" }
                     %>