--- a/app/views/partners/show.html.erb
+++ b/app/views/partners/show.html.erb
@@ -140,7 +140,7 @@
                   <tr>
                     <td><%= dist.created_at.strftime("%m/%d/%Y") %></td>
                     <td><%= dist.storage_location.name %></td>
-                    <td><%= dist.line_items.total %></td>
+                    <td><%= dist.line_items_total %></td>
                     <td class="text-right">
                       <%= view_button_to distribution_path(dist) %>
                       <%= print_button_to print_distribution_path(dist, format: :pdf) %>