--- a/app/views/purchases/_purchase_row.html.erb
+++ b/app/views/purchases/_purchase_row.html.erb
@@ -1,7 +1,7 @@
 <tr>
   <td><%= purchase_row.purchased_from_view %></td>
   <td><%= purchase_row.storage_location.name %></td>
-  <td class="numeric"><%= purchase_row.line_items.total %></td>
+  <td class="numeric"><%= purchase_row.line_items_total %></td>
   <td class="numeric"><%= purchase_row.line_items.size %></td>
   <td class="numeric"><%= dollar_value(purchase_row.amount_spent_in_cents) %></td>
   <td class="date"><%= purchase_row.issued_at.strftime("%F") %></td>