--- a/app/views/transfers/_transfer_row.html.erb
+++ b/app/views/transfers/_transfer_row.html.erb
@@ -3,7 +3,7 @@
   <td><%= transfer_row.to.name %></td>
   <td><%= transfer_row.created_at.strftime("%F") %></td>
   <td><%= transfer_row.comment || "none" %></td>
-  <td><%= transfer_row.line_items.total %></td>
+  <td><%= transfer_row.line_items_total %></td>
   <td>
     <%= delete_button_to transfer_path(transfer_row.id), { confirm: "Are you sure you want to permanently remove this transfer?" } %>
     <%= view_button_to transfer_row %>