--- a/app/views/vendors/show.html.erb
+++ b/app/views/vendors/show.html.erb
@@ -77,7 +77,7 @@
               <% @vendor.purchases.each do |purchase| %>
                 <tr>
                   <td><%= purchase.created_at.strftime("%m/%d/%Y") %></td>
-                  <td><%= purchase.line_items.total %></td>
+                  <td><%= purchase.line_items_total %></td>
                   <td><%= view_button_to purchase, { text: "View purchase details" } %></td>
                 </tr>
               <% end %>