--- a/db/migrate/20210724193104_add_line_items_total_to_distributions.rb
+++ b/db/migrate/20210724193104_add_line_items_total_to_distributions.rb
@@ -1,0 +1,9 @@
+class AddLineItemsTotalToDistributions < ActiveRecord::Migration[6.1]
+  def self.up
+    add_column :distributions, :line_items_total, :integer, null: false, default: 0
+  end
+
+  def self.down
+    remove_column :distributions, :line_items_total
+  end
+end