--- a/db/migrate/20210724193444_add_line_items_total_to_audits.rb
+++ b/db/migrate/20210724193444_add_line_items_total_to_audits.rb
@@ -1,0 +1,9 @@
+class AddLineItemsTotalToAudits < ActiveRecord::Migration[6.1]
+  def self.up
+    add_column :audits, :line_items_total, :integer, null: false, default: 0
+  end
+
+  def self.down
+    remove_column :audits, :line_items_total
+  end
+end