--- a/db/migrate/20210724193454_add_line_items_total_to_donations.rb
+++ b/db/migrate/20210724193454_add_line_items_total_to_donations.rb
@@ -1,0 +1,9 @@
+class AddLineItemsTotalToDonations < ActiveRecord::Migration[6.1]
+  def self.up
+    add_column :donations, :line_items_total, :integer, null: false, default: 0
+  end
+
+  def self.down
+    remove_column :donations, :line_items_total
+  end
+end