--- a/db/migrate/20210724193512_add_line_items_total_to_purchases.rb
+++ b/db/migrate/20210724193512_add_line_items_total_to_purchases.rb
@@ -1,0 +1,9 @@
+class AddLineItemsTotalToPurchases < ActiveRecord::Migration[6.1]
+  def self.up
+    add_column :purchases, :line_items_total, :integer, null: false, default: 0
+  end
+
+  def self.down
+    remove_column :purchases, :line_items_total
+  end
+end