--- a/db/migrate/20210724193520_add_line_items_total_to_transfers.rb
+++ b/db/migrate/20210724193520_add_line_items_total_to_transfers.rb
@@ -1,0 +1,9 @@
+class AddLineItemsTotalToTransfers < ActiveRecord::Migration[6.1]
+  def self.up
+    add_column :transfers, :line_items_total, :integer, null: false, default: 0
+  end
+
+  def self.down
+    remove_column :transfers, :line_items_total
+  end
+end