--- a/db/migrate/20210724194805_add_line_items_total_to_adjustments.rb
+++ b/db/migrate/20210724194805_add_line_items_total_to_adjustments.rb
@@ -1,0 +1,9 @@
+class AddLineItemsTotalToAdjustments < ActiveRecord::Migration[6.1]
+  def self.up
+    add_column :adjustments, :line_items_total, :integer, null: false, default: 0
+  end
+
+  def self.down
+    remove_column :adjustments, :line_items_total
+  end
+end