--- a/db/migrate/20210724194811_add_line_items_total_to_kits.rb
+++ b/db/migrate/20210724194811_add_line_items_total_to_kits.rb
@@ -1,0 +1,9 @@
+class AddLineItemsTotalToKits < ActiveRecord::Migration[6.1]
+  def self.up
+    add_column :kits, :line_items_total, :integer, null: false, default: 0
+  end
+
+  def self.down
+    remove_column :kits, :line_items_total
+  end
+end