--- a/db/migrate/20210724215451_fix_line_item_counts.rb
+++ b/db/migrate/20210724215451_fix_line_item_counts.rb
@@ -1,0 +1,7 @@
+class FixLineItemCounts < ActiveRecord::Migration[6.1]
+  def up
+    LineItem.counter_culture_fix_counts
+  end
+
+  def down; end
+end