--- a/spec/models/distribution_spec.rb
+++ b/spec/models/distribution_spec.rb
@@ -237,7 +237,7 @@ RSpec.describe Distribution, type: :model do
         expect(distribution_details[0]).to eq distribution.partner.name
         expect(distribution_details[1]).to eq distribution.issued_at.strftime("%F")
         expect(distribution_details[2]).to eq distribution.storage_location.name
-        expect(distribution_details[3]).to eq distribution.line_items.total
+        expect(distribution_details[3]).to eq distribution.line_items_total
         expect(distribution_details[5]).to eq distribution.delivery_method
         expect(distribution_details[6]).to eq distribution.state
         expect(distribution_details[7]).to eq distribution.agency_rep