Resolves #2416

### Description

This effectively solves the N+1 query described in the issue, but also introduces a new practice to the project: counter caches.

The implementation is done using the gem [counter_culture](https://github.com/magnusvk/counter_culture), which in my experience is performant, its well documented, and up to date. This gem allow us to add counter caches with more flexibility than the native Rails implementation, for example in this PR a polymorphic quantity total.

For this initial implementation, I'm replacing the instances of `line_items.total` with `line_items_total` that corresponds to the `itemizable` polymorphic parent of the entity using a new column with the same name. This change will require new resources using the `Itemizable` concern to add the column.

I thought about adding a conditional, but in my opinion, it makes sense to have the cache column if we have `line_items`.

Regarding performance, you can see the logs in this gist: "[/diaper_bank/donations logs, old vs new](https://gist.github.com/kinduff/d12942f04fccb697348279f14fcda955)". I'm using a cached environment since it's more realistic for production. But in summary, here is the important part:

**Old**
```
Completed 200 OK in 279ms (Views: 114.6ms | ActiveRecord: 33.7ms | Allocations: 191508)
```

**New**
```
Completed 200 OK in 160ms (Views: 67.6ms | ActiveRecord: 15.2ms | Allocations: 152202)
```

I see a lot of areas we can improve using counter caches, this will without a doubt make the website snappier and consume less resources.

A note about the last commit, I'm adding a migration in order to run the command `LineItem.counter_culture_fix_counts` that will update all the parents with the new total in their corresponding column. I'm not sure if this is something you usually do, I for instance, prefer to run the script in production directly or use data_migrations. Please let me know.

### Type of change

* New feature (non-breaking change which adds functionality)
* This change requires a documentation update

### How Has This Been Tested?

I included the corresponding tests and modified the ones that were failing.

In order to test this, you'll need to pull the changes, migrate, and you're good to go.

Any `line_item` create, update, destroy action should update the column of their parent.

### Selfie

![index](https://user-images.githubusercontent.com/1270156/126881907-914d77be-584b-41aa-8c1c-a2f94c853b3c.gif)