@kinduff 2021-07-25
@edwinthinks I just pushed a fix for the failing spec. 

Regarding counter caches: in my experience if implemented well, and not doing something out of the Rails scope, they work like a charm. Counters are updated inside the transaction of the entity that is being touched. In the case of the test, FactoryBot was triggering them twice from memory.

That being said, if you execute raw SQL, counters are not going to be updated, but neither are callbacks. If you use `update_column`, for example, or `delete`. 

In my opinion, it's worth the trouble, and also for the size of this application and the complexity, it shouldn't cause issues.

I'm all in with this solution and I'm willing to put more time in it if necessary since the 50% performance increase is worth it, but I will respect any decision that the RFG team needs to make.