@edwinthinks 2021-07-25
Hey, @kinduff I appreciate the added context including your experience working in this! I agree with you that if it's done well it works like a charm!  But for a few reasons related to the nature of the project and "business" needs, I'd going to say we should avoid using the counter cache mechanism.

> Regarding counter caches: in my experience if implemented well, and not doing something out of the Rails scope, they work like a charm. Counters are updated inside the transaction of the entity that is being touched. In the case of the test, FactoryBot was triggering them twice from memory.

I believe you 100% that it is implemented well that it works amazingly! However, this project is open source and will hopefully live longer than myself or you being the contributor. I can imagine the chances of overlooking and using `update_column` or even `<<` to trigger an extra callback are pretty great IMO. The mindset that I'am trying to harbor is reducing the number of things newer contributors need to keep in mind and help them avoid chances of grievous issues that could put a negative impact on our users (in this case the line item count being wrong would be potentially devastating).

> In my opinion, it's worth the trouble, and also for the size of this application and the complexity, it shouldn't cause issues.

I also agree with you that at its current application state and team make-up that it may be worth it. However, as mentioned above I want to be cognizant of newer developers that will inherit this project and won't have this context.

> I'm all in with this solution and I'm willing to put more time in it if necessary since the 50% performance increase is worth it, but I will respect any decision that the RFG team needs to make.

In this case, I'd like to voice my disagreement on this. The performance increase is secondary (even tertiary) to the accuracy of the values we are giving as per the "business" needs. For context, our users don't care that it is slow even or have even noticed it. However, they will be very vocal when they notice that the numbers are off and potentially they will make some inaccurate decisions based on them.

I'd like to thank you again for working on this! I really enjoyed learning about counter_culture and possibly use it on a project at work + your selfie haha. Please let me know if you want to discuss any of the points above, happy to give more context and have a dialogue about it!