@kinduff 2021-07-25
I agree with you in some things, in others I do not. My final stand is: if I were the maintainer of this project, I would push the change, but because I'm not, I'm okay with your decision. It's a request anyways.

Hope you can trust this implementation in the future, I'm sure it's going to come around again, counter caches always do, since it's one of the primary performance issues that are fixable this way to avoid N+1 queries.