@edwinthinks 2021-07-25
@kinduff I appreciate your understanding and your flexibility to disagree and commit! 

> I agree with you in some things, in others I do not. 

Love to talk with you sometime about this and discuss our different viewpoints. I certainly don't have the same experience with the library and its application to projects! I'am sure I can learn a ton from you!

Just to be clear are you interested in working on this issue #2444 still but using a different means?